MatchaAI/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.120.4
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.335
langsmith==0.0.92
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.7
llama-index-cli==0.5.3
llama-index-core==0.14.7
llama-index-embeddings-ollama==0.8.3
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.0
llama-index-llms-openai==0.6.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.10.1
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==1.26.4
ollama==0.6.0
openai==1.109.1
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pypdf==6.1.3
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wrapt==1.17.3
yarl==1.22.0
